results/llama_6_turns/scenario_1.5/prompt_4/requirements.txt
Flask==2.2.2
flask_bcrypt==1.0.1
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
PyJWT==2.10.1
python-dotenv==1.2.1
